allow vue-template-compiler transform template into render function
