+++ src/components/presets/List.vue
@@ -1,0 +1,20 @@
+<template>
+  <vue-content-loading v-bind="$attrs" :width="300" :height="120">
+    <rect x="0" y="0" rx="3" ry="3" width="250" height="10" />
+    <rect x="20" y="20" rx="3" ry="3" width="220" height="10" />
+    <rect x="20" y="40" rx="3" ry="3" width="170" height="10" />
+    <rect x="0" y="60" rx="3" ry="3" width="250" height="10" />
+    <rect x="20" y="80" rx="3" ry="3" width="200" height="10" />
+    <rect x="20" y="100" rx="3" ry="3" width="80" height="10" />
+  </vue-content-loading>
+</template>
+
+<script>
+  import VueContentLoading from '../VueContentLoading.vue';
+
+  export default {
+    components: {
+      VueContentLoading,
+    },
+  };
+</script>
